+++ seed/tests/test_dataset.py
@@ -5,10 +5,8 @@ import json
 
 from django.core.urlresolvers import reverse_lazy
 from django.test import TestCase
-
 from superperms.orgs.models import Organization, OrganizationUser
-
-from data_importer.models import ImportFile, ImportRecord
+from seed.data_importer.models import ImportFile, ImportRecord
 from landing.models import SEEDUser as User
