+++ seed/tests/test_models.py
@@ -4,10 +4,8 @@
 from datetime import datetime
 
 from django.test import TestCase
-
 from superperms.orgs.models import Organization, OrganizationUser
-
-from data_importer.models import ImportFile, ImportRecord
+from seed.data_importer.models import ImportFile, ImportRecord
 from landing.models import SEEDUser as User
 from seed import models as seed_models
 from seed.mappings import mapper
