+++ seed/tests/test_project_views.py
@@ -5,9 +5,7 @@ import json
 
 from django.core.urlresolvers import reverse_lazy
 from django.test import TestCase
-
 from superperms.orgs.models import Organization
-
 from landing.models import SEEDUser as User
 from superperms.orgs.models import (
     ROLE_OWNER,
@@ -15,10 +13,9 @@ from superperms.orgs.models import (
     ROLE_VIEWER,
     OrganizationUser,
 )
-
 from seed.tests.util import FakeRequest
 from seed.models import Project
-from data_importer.models import ImportRecord
+from seed.data_importer.models import ImportRecord
 
 
 class ProjectsViewTests(TestCase):
