+++ seed/tests/test_views.py
@@ -10,7 +10,7 @@ from django.core.urlresolvers import reverse_lazy
 from django.test import TestCase
 from superperms.orgs.models import Organization, OrganizationUser
 from seed.audit_logs.models import AuditLog, LOG
-from data_importer.models import ROW_DELIMITER, ImportFile, ImportRecord
+from seed.data_importer.models import ROW_DELIMITER, ImportFile, ImportRecord
 from landing.models import SEEDUser as User
 from seed import decorators
 from seed.factory import SEEDFactory
@@ -36,6 +36,7 @@ from seed.views.main import (
 )
 from seed.utils.mapping import _get_column_names
 from seed.tests import util as test_util
+
